+++ tools/repolint/baseline.json
@@ -9,7 +9,7 @@
     "layering": 1,
     "marker": 0,
     "narrative": 64,
-    "test-file-size": 69938
+    "test-file-size": 69942
   },
   "files": {
     "cmd/e2ebench/main.go": {
@@ -108,7 +108,7 @@
       "test-file-size": 352
     },
     "desktop/frontend/src/__tests__/settings-refresh-snapshot.test.tsx": {
-      "test-file-size": 122
+      "test-file-size": 117
     },
     "desktop/frontend/src/__tests__/tab-switch-hydration.test.tsx": {
       "test-file-size": 162
@@ -141,7 +141,7 @@
       "file-size": 141
     },
     "desktop/frontend/src/components/SettingsPanel.tsx": {
-      "file-size": 6754
+      "file-size": 6718
     },
     "desktop/frontend/src/components/StatusBar.tsx": {
       "file-size": 20
@@ -174,7 +174,7 @@
       "file-size": 269
     },
     "desktop/frontend/src/lib/types.ts": {
-      "file-size": 1456
+      "file-size": 1457
     },
     "desktop/frontend/src/lib/useController.ts": {
       "file-size": 3936
@@ -287,10 +287,10 @@
     "desktop/settings_app.go": {
       "complexity": 22,
       "essay": 26,
-      "file-size": 2795
+      "file-size": 2798
     },
     "desktop/settings_app_test.go": {
-      "test-file-size": 1858
+      "test-file-size": 1865
     },
     "desktop/shared_host.go": {
       "essay": 1
@@ -675,7 +675,7 @@
     "internal/boot/boot.go": {
       "complexity": 298,
       "essay": 102,
-      "file-size": 2241,
+      "file-size": 2242,
       "function-size": 1869,
       "narrative": 3
     },
@@ -977,7 +977,7 @@
     },
     "internal/config/config.go": {
       "essay": 54,
-      "file-size": 1510,
+      "file-size": 1511,
       "narrative": 2
     },
     "internal/config/credentials.go": {
@@ -1012,7 +1012,7 @@
     },
     "internal/config/load.go": {
       "essay": 24,
-      "file-size": 1716,
+      "file-size": 1717,
       "function-size": 53
     },
     "internal/config/main_test.go": {
@@ -1052,13 +1052,13 @@
       "file-size": 267
     },
     "internal/config/render.go": {
-      "complexity": 243,
+      "complexity": 245,
       "essay": 2,
-      "file-size": 989,
-      "function-size": 941
+      "file-size": 995,
+      "function-size": 947
     },
     "internal/config/render_test.go": {
-      "test-file-size": 757
+      "test-file-size": 759
     },
     "internal/config/vision.go": {
       "essay": 1
@@ -1532,7 +1532,7 @@
     "internal/provider/anthropic/anthropic.go": {
       "complexity": 45,
       "essay": 36,
-      "file-size": 118,
+      "file-size": 125,
       "function-size": 154
     },
     "internal/provider/anthropic/anthropic_test.go": {
@@ -1543,10 +1543,10 @@
       "essay": 8
     },
     "internal/provider/openai/openai.go": {
-      "complexity": 67,
+      "complexity": 68,
       "essay": 41,
-      "file-size": 543,
-      "function-size": 263
+      "file-size": 547,
+      "function-size": 267
     },
     "internal/provider/openai/openai_test.go": {
       "essay": 5,
@@ -1562,7 +1562,7 @@
     "internal/provider/responses/responses.go": {
       "complexity": 54,
       "essay": 25,
-      "file-size": 98,
+      "file-size": 105,
       "function-size": 181
     },
     "internal/provider/responses/responses_test.go": {
